ZOOKEEPER-3161: Refactor QuorumPeerMainTest.java: move commonly used functions to base class
Move the following methods to QuorumPeerTestBase.java:

- tearDown()
- LaunchServers()
- waitForOne(), waitForAll()
- logStates()

Author: Andor Molnar <andor@apache.org>

Reviewers: andor@apache.org

Closes #659 from anmolnar/ZOOKEEPER-3161
